+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## [2023.2.12]
 
+### `Project Import` enhancements
+- Auto-registration of the XSD namespace for `editor-definition.xsd` [#804](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/804)
+
 ### `ImpEx` enhancements
 - Added colorization for `odd` and `even` user rights value lines [#802](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/802)